Rename AdventueContextCriteria to ContextRetrievers and update them to create missing contexts for adventures
