+++ Assets/Resources/IncidentData/Character_SummonsDemon_Dies.json
@@ -251,7 +251,7 @@
             "$type": "Game.Incidents.GetOrCreateGreatMonsterAction, Auto-asmdef-SecondPass",
             "criteria": {
               "$id": "26",
-              "$type": "Game.Simulation.MonsterCriteria, Auto-asmdef-SecondPass",
+              "$type": "Game.Simulation.MonsterRetriever, Auto-asmdef-SecondPass",
               "findBySearch": true,
               "monsterData": {
                 "$id": "27",
