+++ Assets/Resources/IncidentData/World_DragonDescendsUponCity.json
@@ -311,7 +311,7 @@
             "$type": "Game.Incidents.GetOrCreateGreatMonsterAction, Auto-asmdef-SecondPass",
             "criteria": {
               "$id": "28",
-              "$type": "Game.Simulation.MonsterCriteria, Auto-asmdef-SecondPass",
+              "$type": "Game.Simulation.MonsterRetriever, Auto-asmdef-SecondPass",
               "findBySearch": true,
               "monsterData": null,
               "isLegendary": true,
