--- Assets/Resources/ScriptableObjects/Encounters/Monster_Lair_00.asset
@@ -1,228 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 084bf99f985f0d341a9f6e9f9a93e27c, type: 3}
-  m_Name: Monster_Lair_00
-  m_EditorClassIdentifier: 
-  serializationData:
-    SerializedFormat: 2
-    SerializedBytes: 
-    ReferencedUnityObjects: []
-    SerializedBytesString: 
-    Prefab: {fileID: 0}
-    PrefabModificationsReferencedUnityObjects: []
-    PrefabModifications: []
-    SerializationNodes:
-    - Name: contextCriterium
-      Entry: 7
-      Data: 0|System.Collections.Generic.List`1[[Game.Simulation.IAdventureContextCriteria,
-        Auto-asmdef-SecondPass]], mscorlib
-    - Name: 
-      Entry: 12
-      Data: 2
-    - Name: 
-      Entry: 7
-      Data: 1|Game.Simulation.MonsterCriteria, Auto-asmdef-SecondPass
-    - Name: CriteriaID
-      Entry: 3
-      Data: 0
-    - Name: contextTypeName
-      Entry: 1
-      Data: Monster
-    - Name: historical
-      Entry: 5
-      Data: false
-    - Name: findBySearch
-      Entry: 5
-      Data: false
-    - Name: monsterData
-      Entry: 7
-      Data: 2|Game.Incidents.ScriptableObjectRetriever`1[[Game.Data.MonsterData,
-        Auto-asmdef-SecondPass]], Auto-asmdef-SecondPass
-    - Name: prefabKey
-      Entry: 1
-      Data: Worg
-    - Name: onChanged
-      Entry: 6
-      Data: 
-    - Name: 
-      Entry: 8
-      Data: 
-    - Name: isLegendary
-      Entry: 5
-      Data: false
-    - Name: isLandDwelling
-      Entry: 5
-      Data: true
-    - Name: allowedSizes
-      Entry: 7
-      Data: 3|System.Collections.Generic.List`1[[Game.Enums.CreatureSize, Auto-asmdef-SecondPass]],
-        mscorlib
-    - Name: 
-      Entry: 12
-      Data: 2
-    - Name: 
-      Entry: 3
-      Data: 1
-    - Name: 
-      Entry: 3
-      Data: 2
-    - Name: 
-      Entry: 13
-      Data: 
-    - Name: 
-      Entry: 8
-      Data: 
-    - Name: allowedTypes
-      Entry: 7
-      Data: 4|System.Collections.Generic.List`1[[Game.Enums.CreatureType, Auto-asmdef-SecondPass]],
-        mscorlib
-    - Name: 
-      Entry: 12
-      Data: 2
-    - Name: 
-      Entry: 3
-      Data: 1
-    - Name: 
-      Entry: 3
-      Data: 9
-    - Name: 
-      Entry: 13
-      Data: 
-    - Name: 
-      Entry: 8
-      Data: 
-    - Name: allowedAlignments
-      Entry: 7
-      Data: 5|System.Collections.Generic.List`1[[Game.Enums.CreatureAlignment, Auto-asmdef-SecondPass]],
-        mscorlib
-    - Name: 
-      Entry: 12
-      Data: 5
-    - Name: 
-      Entry: 3
-      Data: 2
-    - Name: 
-      Entry: 3
-      Data: 4
-    - Name: 
-      Entry: 3
-      Data: 5
-    - Name: 
-      Entry: 3
-      Data: 7
-    - Name: 
-      Entry: 3
-      Data: 8
-    - Name: 
-      Entry: 13
-      Data: 
-    - Name: 
-      Entry: 8
-      Data: 
-    - Name: 
-      Entry: 8
-      Data: 
-    - Name: 
-      Entry: 7
-      Data: 6|Game.Simulation.AdventureCharacterContextCriteria, Auto-asmdef-SecondPass
-    - Name: CriteriaID
-      Entry: 3
-      Data: 0
-    - Name: contextTypeName
-      Entry: 1
-      Data: Character
-    - Name: historical
-      Entry: 5
-      Data: false
-    - Name: 
-      Entry: 8
-      Data: 
-    - Name: 
-      Entry: 13
-      Data: 
-    - Name: 
-      Entry: 8
-      Data: 
-    - Name: components
-      Entry: 7
-      Data: 7|System.Collections.Generic.List`1[[Game.Simulation.IAdventureComponent,
-        Auto-asmdef-SecondPass]], mscorlib
-    - Name: 
-      Entry: 12
-      Data: 4
-    - Name: 
-      Entry: 7
-      Data: 8|Game.Simulation.AdventureNarrationComponent, Auto-asmdef-SecondPass
-    - Name: ComponentID
-      Entry: 3
-      Data: 0
-    - Name: text
-      Entry: 6
-      Data: 
-    - Name: 
-      Entry: 8
-      Data: 
-    - Name: 
-      Entry: 7
-      Data: 9|Game.Simulation.AdventureTextComponent, Auto-asmdef-SecondPass
-    - Name: ComponentID
-      Entry: 3
-      Data: 0
-    - Name: text
-      Entry: 6
-      Data: 
-    - Name: 
-      Entry: 8
-      Data: 
-    - Name: 
-      Entry: 7
-      Data: 10|Game.Simulation.AdventureTextComponent, Auto-asmdef-SecondPass
-    - Name: ComponentID
-      Entry: 3
-      Data: 0
-    - Name: text
-      Entry: 6
-      Data: 
-    - Name: 
-      Entry: 8
-      Data: 
-    - Name: 
-      Entry: 7
-      Data: 11|Game.Simulation.AdventureNarrationComponent, Auto-asmdef-SecondPass
-    - Name: ComponentID
-      Entry: 3
-      Data: 0
-    - Name: text
-      Entry: 6
-      Data: 
-    - Name: 
-      Entry: 8
-      Data: 
-    - Name: 
-      Entry: 13
-      Data: 
-    - Name: 
-      Entry: 8
-      Data: 
-  encounterTitle: A Lair of Monsters
-  encounterDifficulty: 9
-  majorEncounter: 0
-  encounterLocationType: 0
-  encounterTypes: 00000000
-  allowedBiomes: 0000000001000000020000000300000005000000
-  encounterBlurb: '{1} has called upon a group of adventurers to remove a group of
-    {0} that have been terrorizing the nearby countryside.'
-  encounterSummary: '{1} explains that the monsters, a -0- of {0}s, have been attacking
-    and killing the town''s livestock, causing a shortage of food and putting the
-    town in danger of starvation. [1] offers a reward to the adventurers if they
-    can successfully remove the {0}s from the area. From information gathered from
-    local woodsman, [1] believes that the {0}s can be found near {LOCATION}.'
