+++ Assets/Scripts/Simulation/Adventures/AdventureComponents/AdventureCombatComponent.cs
@@ -28,7 +28,7 @@ namespace Game.Simulation
 #if UNITY_EDITOR
 		private IEnumerable<int> GetCombatantIDs()
 		{
-			var ids = EncounterEditorWindow.contextCriterium.Where(x => x.GetType() == typeof(MonsterCriteria)).Select(x => x.CriteriaID);
+			var ids = EncounterEditorWindow.contextCriterium.Where(x => x.GetType() == typeof(MonsterRetriever)).Select(x => x.RetrieverID);
 			return ids;
 		}
 #endif
