--- Assets/Scripts/Simulation/Adventures/AdventureContextCriteria.cs
@@ -1,81 +1,0 @@
-﻿using Game.Incidents;
-using Newtonsoft.Json;
-using Sirenix.OdinInspector;
-using System;
-using System.Collections.Generic;
-using System.Text.RegularExpressions;
-using UnityEngine;
-
-namespace Game.Simulation
-{
-	[Serializable, HideReferenceObjectPicker]
-	abstract public class AdventureContextCriteria<T> : IAdventureContextCriteria where T : IIncidentContext
-	{
-		public Type ContextType => typeof(T);
-
-		[JsonIgnore]
-		public IIncidentContext Context
-		{
-			get
-			{
-				if(context == null)
-				{
-					RetrieveContext();
-				}
-
-				return context;
-			}
-			set
-			{
-				context = value;
-			}
-		}
-
-		[JsonIgnore]
-		public T TypedContext => (T)Context;
-
-		private IIncidentContext context;
-		public bool IsHistorical => historical;
-
-		[SerializeField, ReadOnly, HorizontalGroup(LabelWidth = 120), PropertyOrder(-1)]
-		private string contextTypeName;
-		[SerializeField, HorizontalGroup, PropertyOrder(-1), ReadOnly]
-		public int CriteriaID { get; set; }
-		[SerializeField, HorizontalGroup, PropertyOrder(-1)]
-		public bool historical;
-
-		abstract public Dictionary<string, Func<T, int, string>> Replacements { get; }
-		abstract public void SpawnPopup();
-
-		public AdventureContextCriteria()
-		{
-			contextTypeName = ContextType.Name;
-		}
-
-		protected T GetTypedContext()
-		{
-			return (T)Context;
-		}
-
-		abstract public void RetrieveContext();
-		public void ReplaceTextPlaceholders(ref string text)
-		{
-			if(string.IsNullOrEmpty(text))
-			{
-				return;
-			}
-
-			foreach(var pair in Replacements)
-			{
-				var idReplacementPattern = $"{CriteriaID}";
-				var toReplace = Regex.Replace(pair.Key, "##", idReplacementPattern);
-				var replaceWith = pair.Value(GetTypedContext(), CriteriaID);
-				text = text.Replace(@toReplace, replaceWith);
-			}
-
-			//For capitalizing after periods
-			var postPeriod = new Regex(@"(\. )([a-z])");
-			text = postPeriod.Replace(text, m => m.Groups[1].Value + m.Groups[2].Value.ToUpper());
-		}
-	}
-}
