--- Assets/Scripts/Simulation/Adventures/AdventureContextCriteria.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 110a29f40be042d498a20c94370cea8b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
