--- Assets/Scripts/Simulation/Adventures/IAdventureContextCriteria.cs
@@ -1,17 +1,0 @@
-﻿using Game.Incidents;
-using System;
-using System.Collections.Generic;
-
-namespace Game.Simulation
-{
-	public interface IAdventureContextCriteria 
-	{
-		public Type ContextType { get; }
-		public int CriteriaID { get; set; }
-		public IIncidentContext Context { get; set; }
-		public bool IsHistorical { get; }
-
-		public void ReplaceTextPlaceholders(ref string text);
-		public void SpawnPopup();
-	}
-}
