--- Assets/Scripts/Simulation/Adventures/MonsterCriteria.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8a3490b3822d3254cbeb73685112e72e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
