--- Assets/Scripts/Simulation/Adventures/TreasureCriteria.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 56f12f207c76de3488ca73ebff5f2a9c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
